Fixed an off-by-one error selection error.
If the user tried to select text when there was no label nearby, they
would be unable to select the last character of a text node. This has
been fixed.

Note that this would happen essentially when labels are present but
not displayed. Not a common case with the generic mode.