Prevent focus recovery from scrolling the editor.
A workaround for a bug in FF causes issues in all other
browsers. Basically, when the editor loses focus while the editor pane
is scrolled to *not* show the caret and the focus is regained, the
editor pane would scroll to show the caret. The new code prevents
this. There is still some cases where it can happen in FF but these
are **extremely** hard to reproduce by hand.