Fixed a focusing problem.
If something had scrolled the editor pane between the loss of focus
and our regaining it, the scroll position would get lost. In
particular, this happened in Chrome when the user used the search
function to search text on the page. Chrome would scroll the page to
the hit. Without keeping the scroll intact, then the user would lose
their position when they clicked into the window to (for instance)
select the hit.