Enable wed to operate with transient carets.
The changes to support attributes have violated assumptions that held
true up to now:

1. Validating the document does not affect the caret. This became false
due to possible attribute errors, which when added to the GUI tree would
corrupt the caret.

2. Refreshing an element does not affect the caret. This became false as
start labels contain attributes which may contain the caret. Refreshing
an element *will* corrupt the caret if it was in one of the element's
attributes.

To counter this, the following changes were made:

1. Errors are displayed in an error layer.

2. A "scroller" element scrolls the editing pane and the error layer.

3. The default decorator saves and restore caret positions.