Fixed a bug in the build process.
The process used to rely on the time stamps of files inside NPM
packages. This is brittle because when NPM installs a package, it
preserve the time stamps of the package. This could lead to cases where
installing a new package would not trigger the build process to copy the
new package files!