Don't use $(shell find...) as dependency.
You don't want to have $(shell find...) as a dependency when the list of
files is going to change dynamically during the build. The issue that
required this change is that when upgrading from lodash 2.x to 3.x some
files disappeared. When $(shell find...) executed the files were
present, but as the build progressed, the files of 2.x were replaced
with 3.x and by the time make got to the rule with $(shell find...) some
files that were initially present were missing.

We use a shell check instead to see whether we really need to optimize
again.