Fixed context menus on touch devices.
The march of progress made it so that Chrome is now better able to
detect whether touch events are available. This, in turn, causes
Bootstrap to assume it is on a mobile platform whenever touch events are
available. This causes Bootstrap to add a backdrop to capture clicks
outside dropdowns, which causes problems with our context menus. The
problem has been fixed.