Split the two interfaces covered by this polyfill.
The polyfill used to mix the tests and patching of two different
interfaces: ParentNode and NonDocumentTypeChildNode. This would yield
incorrect results sometimes, because support for one interface does not
entail support for the other. This commit fixes this. The polyfill takes
care of both interfaces but disentangle the tests and the patching to
make it so that once the file has run both interfaces are supported on
HTML and XML nodes.