Fixed an IE11 inconsistency regarding attributes.
There was an inconsistency between IE11 and other browsers in the way
deletion of attributes was handled. When an attribute is deleted, the
caret is put in the "next" attribute. IE11 disagreed with other browsers
as to which attribute was next in the data tree. This has been fixed by
relying on the GUI tree.