Update the way we test for attributes.
This was prompted by the move to jsdom 8.x, which no longer has
attributes inherit the ``Node`` interface.

This is needed because wed works with HTML and XML DOM trees and
unfortunately, things have gotten murky. Once upon a time, an attribute
was determined by checking the ``nodeType`` field. This worked both for
HTML and XML nodes. This worked because attributes inherited from
``Node``, which is the DOM interface that defines ``nodeType``. It was
paradise.

Then the luminaries that drive the implementation of DOM in actual
browsers decided that attributes were no longer really nodes. So they
decided to make attribute objects inherit from the ``Attr``
interface **only**. This means that ``nodeType`` no longer exists for
attributes. The new way to test whether something is an attribute is to
test with ``instanceof Attr``. However, as usual, the DOM implementation
for XML lags behind the HTML side and on Chrome 49 (to name just one
case), ``instanceof Attr`` does not work on XML attributes whereas
testing ``nodeType`` does.