Test onbeforeunload directly.
We no longer perform the reload tests through Selenium. Around the time
of Chrome 52-53, Selenium started hanging on driver.refresh() if
onbeforeunload showed a dialog box. This seems to be a bug in Chrome or
chromedriver or something. Rather than chase the bug, we've dropped
these tests here and instead check in wed_test.js that onbeforeunload
returns expected values. This won't trap issues with bizarro platforms,
unfortunately but that was the expedient thing to do.