Fix a bug in resizing.
There was a problem that could happen if a user was resizing a modal
beyond its maximum size. The mouse pointer could move outside the
modal. When this happened, it resulted in a mouseup on the backdrop that
Bootstrap puts behind the modal to detect clicks outside of it, and it
would close the modal when the user stopped resizing it. The fix makes
it so that the modal is no longer closed due to this problem.