Upgrade log4javascript and add log.clearAppenders.
We upgraded log4javascript. This has revealed a problem with log.js. In
testing, wed could keep appending ajax appenders without ever cleaning
them up, which caused a progressive slowdown. The new update amplified
the problem (because it fixed a bug that hid it previously). So we added
a clearAppenders method to clear the appenders between tests.