Upgrade bootbox, fix its configuration.
We switched to the package "bootbox". It turns out that "bootbox.js" had
been silently deprecated and was no longer up to date. Also discovered
that bootbox has become an AMD module and the shim was usless. However,
it needs a special map to handle the fact that it depends on both
bootstrap and jquery but only lists jquery as a dependency.