Turn off allowUncaught.
Allow uncaught was turned on early on in the test suite's existence.  It was
used to handle cases that *probably* are no longer an issue.  We turned it back
off because it has a severe consequence that non-terminating or failing promises
are not detectable when turned on:

https://github.com/mochajs/mocha/issues/2281

If somehow we need this back on, we'll have to fix the problem above.