Fixed problem with DELETE key on labels.
The recent changes to ``domutil`` revealed a long-existing bug in the code that
handles DELETE when issued on an element label. It should have been calling
``terminate()`` from the get go. In the past, the lack of termination turned out
to not be a problem because ``domutil.nextCaretPosition`` would actually return
something for a caret that was no longer in the tree. The position returned was
such that the code that used it would do a no-op anyway. So the bug had no
visible effect.

The recent changes to ``domutil`` make it return ``null`` instead of returning a
position. This lead to a crash which is fixed by making sure that the code
handling DELETE on labels ends right away.