Use BeforeInsertNodeAt to preprocess new nodes.
The old code used the events provided by updater_domlistener, which made it so
that the DOM elements in the GUI tree would be updated *after* they are in the
DOM tree. This caused quite a bit of a slowdown. We now use BeforeInsertNodeAt
to process them *before* they get into the tree.