Simplify caret refresh algorithm.
The old algorithm would insert a marker element into the DOM, take its position
and height, and then remove it. This required extremely ad hoc code and caused
some complications (like sometimes having to restore a damaged selection). The
new code uses ``boundaryXY`` and gets a line height from querying the DOM. So it
does not modify the DOM tree. ``boundaryXY`` is a fairly complex piece of code,
but since it exists for use elsewhere... we can just use it here too and
simplify the code-base overall.