Fix some scrolling issues.
The recent simplification of the test code revealed that there was a bug in the
scrolling code. In particular focusing the input field would scroll the window
back to the top, and this causes problems if the window was scrolled down. So we
no longer move the input field off-screen. We also take scrollbars into account
when scrolling the contents of GUI tree.