Drop xml-to-html and html-to-xml entirely.
They had been retired from public view a long time ago but were still used in
the test suite for some tests because some tests needed operating on the HTML
converted tree but jsdom did not support DOMParser so the HTML tree had to be
constructed outside the test itself.

With the upgrade to jsdom 11 we now have support for DOMParser so the whole
shebang is no longer necessary.