Move caret layer into scroller and extend beyond visible document.
Previously, the caret layer was outside the scroller and designed to be only
just as big as the displayed portion of the document. One major consequence was
that the caret marker had to be updated by custom JavaScript code with each
scroll of the document, to move it.

The caret layer is now inside the scroller and is automatically scrolled with
the document. Therefore the caret marker does not need updating with every
scroll of the document.

This paves the way to easily add other items like highights and selection
markers in the caret layer.