Rework test tasks to prevent test timeouts.
The previous setup would allow the base Karma tests, and the Karma tests for
Webpack to launch in parallel, which could cause failures in the base Karma
tests due to CPU exhaustion and random timeouts in Mocha. (Webpack is very
CPU-intensive). The new code requires that the tests run sequentially, and thus
prevents the timeouts.